+++ js/workflows/telegram-capture/03_compose-response__7a117ec5-8af2-4d89-a227-128c2c488114.js
@@ -7,12 +7,12 @@
  */
 'use strict';
 
-const { getConfigStatic } = require('../../../src/libs/config.js');
+const { getConfig } = require('../../../src/libs/config.js');
 
 module.exports = async function run(ctx) {
   const { $input, $json, $items, $node, $env, helpers } = ctx;
 
-const config = getConfigStatic();
+const config = getConfig();
 const isTestMode = !!(config && config.db && config.db.is_test_mode);